+++ examples/standalone/todomvc/src/livestore/schema.ts
@@ -11,7 +11,6 @@ const todos = DbSchema.table(
     text: DbSchema.text({ default: '' }),
     completed: DbSchema.boolean({ default: false }),
     deleted: DbSchema.integer({ nullable: true, schema: Schema.DateFromNumber }),
-    importantField: DbSchema.text({}),
   },
   { deriveMutations: true },
 )